+++ Changelog.md
@@ -7,6 +7,7 @@
 - Fix contention.
 - Micro optimization during first call of `ObjectPool<T>.Shared.Return(T)` in an specific thread.
 - Micro optimization during automatic trimming of `ObjectPool<T>.Shared`.
+- Support `Trim(bool)` in `ObjectPool<T>.Shared`.
 
 ## v0.2.3
 - Improve documentation.